+++ EDDI/ChangeLog.md
@@ -28,6 +28,8 @@ Full details of the variables available for each noted event, and VoiceAttack in
       * `Carrier bank transfer` added.
       * `Carrier decommission cancelled` added.
       * `Carrier decommission scheduled` added.
+      * `Carrier docking permission` added.
+      * `Carrier fuel deposit` added.
       * `Carrier purchased` added, triggered when you purchase a fleet carrier.
       * `Carrier service changed` added, triggered when you change the services available at your fleet carrier.
       * `Carrier stats` added, triggered when you open the carrier management screen.